+++ server/src/weaverbird/backends/pypika_translator/translators/base.py
@@ -344,7 +344,7 @@ class SQLTranslator(ABC):
             raise ValueError("No steps provided")
         assert (
             steps[0].name == "domain" or steps[0].name == "customsql"
-        ), "First step must be one of domain or customsql step"
+        ), f"First step must be one of domain or customsql step, got '{steps[0].name}'"
         self._step_count = 0
 
         # A single custom SQL step must always be wrapped in a CTE, as we cannot apply offset and